format everything with scalafmt
I was kinda wary of enabling scalafmt for the whole project
in scalameta 1.x, because large parts of our codebase were written
before scalafmt existed.

However with the new beginnings in this project, we can start
well-formatted and then make sure things stay well-formatted
as we're reviving more and more scalameta 1.x modules.