first stab at scala.macros
In a crucial UX improvement over scala.reflect.macros, we implement
the scala.macros.Universe cake in the scala.macros package object.

In the previous commit, we modelled all companions using defs, not vals,
and this proves to be critical, making it possible to swap universe
implementations on the fly (see the sources of the package object).

In comparison with liufengyun/gestalt, in order to implement Universe,
we only need to implement one method. We also need just one asInstanceOf
to trick the scalac/dotc typechecker.