separate abiDef from implDef
Sometimes we may want implDef to be called in a particular way.

For instance, old-style macros trim stacktraces of macro-generated
exceptions by looking for a method whose name ends with
`macroExpandWithRuntime` (I remember being criticized for this particular
way of arranging the macro engine, and now I understand why I was wrong).

Thus, for optimal user experience on platforms that desugar new-style
macros to old-style macros, we really want implDef to be named in this
idiosyncratic way.

However, we also want the inline module to have a method with
a well-defined name that can be called in platform-independent fashion.
That's why we introduce the notion of abiDef.