initial semantic API
Extends the cake with a semantic slice that features the usual structure.
This design tries to combine the best features of the scalameta semantic
API ca. 2015, the scalameta semantic API ca. 2017 and liufengyun/gestalt.

This commit introduces Symbol and Denotation, which are both opaque
entities fully defined by a concrete implementation. A Symbol is a GUID
of a member (where uniqueness is defined within the limits of the
definition of classpath). A Denotation is a Symbol augmented with
a prefix, which enables it to compute symbol signatures.

Types from syntactic API are also used to represent types in the
semantic API. Fengyun was insisting that we shouldn't conflate these
two concepts, but I'd like to use this pull request as an extended
illustration of the idea that I tried to convey. Hopefully, this and
the follow-up commits will provide arguments in favor of this approach.

Semantic operations are implemented in an unusual fashion. There are a
lot of methods that overlap between refs, members, symbols and denots,
because e.g. it is reasonable to expect isFinal to be callable on all of
the above. As a result, quite some design effort went into efficiently
sharing these operations without copy/paste. I'm pretty happy with the
end result.

One blatant issue with the current data model is that allows performing
semantic operations on unattributed trees. There is a solution for that
that Fengyun has implemented in liufengyun/gestalt (separation of
typed and untyped trees), and it works really well. Therefore, we don't
consider this issue to be a serious problem, and we'll get to fixing it
once more urgent tasks are dealt with.