introduce Mirror
Even though we went back to the reflect-like scheme of organizing things,
we shouldn't regress to complete savagery of semantic APIs being powered
by global state of the cake.

This commit also includes a reorganization of expansions, because
I realized that there's no need to have any method in Expansion
if it lives inside the cake. This saves us an additional layer of
abstraction.

I've also applied aggressive abbreviation, replacing all reasonable
occurrences of `symbol` with just `sym`. In the world of `tpe` and `denot`
this seems logical. Let's see how people like it.