replace inline/meta with def/macro
As the first step for the new macro system, we want to get just the
macro-related functionality, and for that we don't have to implement the
full power of inline and meta.

Since we don't need to implement inline and meta separately right away,
inline/meta becomes just a fancy way of indicating a new-style macro,
and this syntax starts looking a bit excessive.

Instead of introducing two new keywords and bringing our users the pain
of migrating code that uses these keywords as identifiers, we propose
that for the time being we use the old def/macro syntax with the
block-shaped right-hand side as the syntax for new-style macros.