Remove unused quasiquotes implementation.
Quasiquotes have undoubtedly proven themselvese to be a huge
enabler for macro authors. However, quasiquotes also slow us down while
we experiment in the new macros. In addition, Squid shows us that it's
possible to implement reusable, typesafe, hygienic and well-scoped
quasiquotes as an extension on top of an existing macro system. For this
reason, we remove quasiquotes from the core macro engine, simplifying
our infrastructure.